requirements.txt
Flask
python-dotenv
gpt4all
python-telegram-bot==13.15
discord.py
sqlalchemy
